LayoutTests/fast/events/message-port-deleted-document-expected.txt

CONSOLE MESSAGE: line 29: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that destroying a document doesn't cause a crash when posting a message to a MessagePort it owned.
